Refactor test, use junit 5
