+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symbols/AnnotationReflectionTest.java
@@ -10,65 +10,63 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.lang.reflect.Parameter;
 import java.util.List;
 
-import org.checkerframework.checker.nullness.qual.NonNull;
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+import org.pcollections.PSet;
 
 import net.sourceforge.pmd.lang.java.JavaParsingHelper;
+import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclarator;
 import net.sourceforge.pmd.lang.java.symbols.SymbolicValue.SymAnnot;
-import net.sourceforge.pmd.lang.java.symbols.internal.asm.AsmSymbolResolver;
+import net.sourceforge.pmd.lang.java.symbols.internal.SymImplementation;
 import net.sourceforge.pmd.lang.java.symbols.testdata.AnnotWithDefaults;
+import net.sourceforge.pmd.lang.java.symbols.testdata.LocalVarAnnotation;
 import net.sourceforge.pmd.lang.java.symbols.testdata.SomeClass;
-import net.sourceforge.pmd.lang.java.types.JPrimitiveType.PrimitiveTypeKind;
-import net.sourceforge.pmd.lang.java.types.JTypeMirror;
-import net.sourceforge.pmd.lang.java.types.Substitution;
 import net.sourceforge.pmd.lang.java.types.TypeSystem;
 
-public abstract class AbstractSymbolTest {
+public class AnnotationReflectionTest {
 
-    protected final TypeSystem ts = JavaParsingHelper.TEST_TYPE_SYSTEM;
-    protected final AsmSymbolResolver loader = (AsmSymbolResolver) ts.bootstrapResolver();
-    
-    private final boolean expectDebugSymbols;
-    
-    public AbstractSymbolTest(boolean expectDebugSymbols) {
-        this.expectDebugSymbols = expectDebugSymbols;
+    private final TypeSystem ts = JavaParsingHelper.TEST_TYPE_SYSTEM;
+
+    public AnnotationReflectionTest() {
     }
 
-    protected abstract JClassSymbol resolveSymbol(Class<?> clazz);
 
-    @Test
-    public void testReflectionOfClassMethods() {
+    @ParameterizedTest
+    @EnumSource
+    public void testReflectionOfClassMethods(SymImplementation impl) {
         Class<SomeClass> actualClass = SomeClass.class;
-        JClassSymbol sym = resolveSymbol(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
     
         Assert.assertNotNull(sym);
-        assertAllMethodsMatch(actualClass, sym);
-        assertAllFieldsMatch(actualClass, sym);
+        impl.assertAllMethodsMatch(actualClass, sym);
+        impl.assertAllFieldsMatch(actualClass, sym);
     }
 
-    @Test
-    public void testReflectionOfAnnotDefaults() {
+    @ParameterizedTest
+    @EnumSource
+    public void testReflectionOfAnnotDefaults(SymImplementation impl) {
         Class<AnnotWithDefaults> actualClass = AnnotWithDefaults.class;
-        JClassSymbol sym = loadAnnotation(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
         
         Assert.assertNotNull(sym);
-        assertAllMethodsMatch(actualClass, sym);
+        impl.assertAllMethodsMatch(actualClass, sym);
     }
-    
-    @Test
-    public void testRetentionClassAnnot() {
+
+    @ParameterizedTest
+    @EnumSource
+    public void testRetentionClassAnnot(SymImplementation impl) {
         // note that as the annotation has retention class, we can't use reflection to check
 
         /*
             @AnnotWithDefaults(valueNoDefault = "ohio",
                        stringArrayDefault = {})
          */
-        JClassSymbol sym = resolveSymbol(SomeClass.class);
+        JClassSymbol sym = impl.getSymbol(ts, SomeClass.class);
 
         Assert.assertTrue(sym.isAnnotationPresent(AnnotWithDefaults.class));
 
@@ -88,30 +86,33 @@ public abstract class AbstractSymbolTest {
         // Non existing values are always considered unknown
         Assert.assertEquals(UNKNOWN, target.attributeMatches("attributeDoesNotExist", new String[] {"a"}));
     }
-    
-    @Test
-    public void testAnnotOnType() {
+
+    @ParameterizedTest
+    @EnumSource
+    public void testAnnotOnType(SymImplementation impl) {
         // This only checks for Target.TYPE annotations
         Class<AnnotWithDefaults> actualClass = AnnotWithDefaults.class;
-        JClassSymbol sym = resolveSymbol(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
 
         assertEqualsAnnotations(actualClass, sym);
     }
-    
-    @Test
-    public void testAnnotOnAnnot() {
+
+    @ParameterizedTest
+    @EnumSource
+    public void testAnnotOnAnnot(SymImplementation impl) {
      // This only checks for Target.ANNOTATION_TYPE annotations
         Class<AnnotWithDefaults> actualClass = AnnotWithDefaults.class;
-        JClassSymbol sym = resolveSymbol(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
 
         assertEqualsAnnotations(actualClass, sym);
     }
-    
-    @Test
-    public void testAnnotOnParameter() {
+
+    @ParameterizedTest
+    @EnumSource
+    public void testAnnotOnParameter(SymImplementation impl) {
         // This only checks Target.PARAMETER annotations, do not confuse with TYPE_PARAMETER / TYPE_USE
         Class<SomeClass> actualClass = SomeClass.class;
-        JClassSymbol sym = resolveSymbol(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
 
         List<JMethodSymbol> ms = sym.getDeclaredMethods();
         Method[] actualMethods = actualClass.getDeclaredMethods();
@@ -132,11 +133,12 @@ public abstract class AbstractSymbolTest {
         }
     }
 
-    @Test
-    public void testAnnotOnField() {
+    @ParameterizedTest
+    @EnumSource
+    public void testAnnotOnField(SymImplementation impl) {
         // This only checks Target.FIELD annotations, do not confuse with TYPE_USE annotations
         Class<SomeClass> actualClass = SomeClass.class;
-        JClassSymbol sym = resolveSymbol(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
         
         List<JFieldSymbol> fs = sym.getDeclaredFields();
         Field[] actualFields = actualClass.getDeclaredFields();
@@ -149,11 +151,12 @@ public abstract class AbstractSymbolTest {
         }
     }
     
-    @Test
-    public void testAnnotOnMethod() {
+    @ParameterizedTest
+    @EnumSource
+    public void testAnnotOnMethod(SymImplementation impl) {
         // This only checks Target.METHOD annotations, do not confuse with TYPE_USE on return types
         Class<SomeClass> actualClass = SomeClass.class;
-        JClassSymbol sym = resolveSymbol(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
 
         List<JMethodSymbol> ms = sym.getDeclaredMethods();
         Method[] actualMethods = actualClass.getDeclaredMethods();
@@ -166,11 +169,12 @@ public abstract class AbstractSymbolTest {
         }
     }
     
-    @Test
-    public void testAnnotOnConstructor() {
+    @ParameterizedTest
+    @EnumSource
+    public void testAnnotOnConstructor(SymImplementation impl) {
      // This only checks Target.CONSTRUCTOR annotations, do not confuse with TYPE_USE on return types
         Class<SomeClass> actualClass = SomeClass.class;
-        JClassSymbol sym = resolveSymbol(actualClass);
+        JClassSymbol sym = impl.getSymbol(ts, actualClass);
 
         List<JConstructorSymbol> cs = sym.getConstructors();
         Constructor<?>[] actualConstructors = actualClass.getDeclaredConstructors();
@@ -183,83 +187,28 @@ public abstract class AbstractSymbolTest {
             assertEqualsAnnotations(c, mSym);
         }
     }
+
+    @Test
+    public void testAnnotOnLocalVar() {
+        // This only checks Target.LOCAL_VAR annotations, do not confuse with TYPE_USE on return types
+        JClassSymbol sym = SymImplementation.AST.getSymbol(ts, SomeClass.class);
+
+        JMethodSymbol method = sym.getDeclaredMethods().stream().filter(it -> it.getSimpleName().equals("withAnnotatedLocal"))
+                                  .findFirst().orElseThrow(AssertionError::new);
+
+        ASTVariableDeclarator variableDeclarator = method.tryGetNode().descendants(ASTVariableDeclarator.class).first();
+        JVariableSymbol localSym = variableDeclarator.getSymbolTable().variables().resolveFirst("local").getSymbol();
+
+        PSet<SymAnnot> declaredAnnotations = localSym.getDeclaredAnnotations();
+
+        Assert.assertEquals(1, declaredAnnotations.size());
+        Assert.assertNotNull(localSym.getDeclaredAnnotation(LocalVarAnnotation.class));
+    }
     
     protected SymbolicValue symValueOf(Object o) {
         return SymbolicValue.of(ts, o);
     }
-    
-    protected @NonNull JClassSymbol loadAnnotation(Class<?> klass) {
-        JClassSymbol sym = resolveSymbol(klass);
-        Assert.assertTrue(sym.isAbstract());
-        Assert.assertTrue(sym.isAnnotation());
-        Assert.assertTrue(sym.isInterface());
-        return sym;
-    }
 
-    private void assertAllFieldsMatch(Class<?> actualClass, JClassSymbol sym) {
-        List<JFieldSymbol> fs = sym.getDeclaredFields();
-        Field[] actualFields = actualClass.getDeclaredFields();
-        Assert.assertEquals(actualFields.length, fs.size());
-   
-        for (final Field f : actualFields) {
-            JFieldSymbol fSym = fs.stream().filter(it -> it.getSimpleName().equals(f.getName()))
-                    .findFirst().orElseThrow(AssertionError::new);
-            
-            // Type matches
-            final JTypeMirror expectedType = typeMirrorOf(f.getType());
-            Assert.assertEquals(expectedType, fSym.getTypeMirror(Substitution.EMPTY));
-            
-            Assert.assertEquals(f.getModifiers(), fSym.getModifiers());
-        }
-    }
-    
-    private void assertAllMethodsMatch(Class<?> actualClass, JClassSymbol sym) {
-        List<JMethodSymbol> ms = sym.getDeclaredMethods();
-        Method[] actualMethods = actualClass.getDeclaredMethods();
-        Assert.assertEquals(actualMethods.length, ms.size());
-   
-        for (final Method m : actualMethods) {
-            JMethodSymbol mSym = ms.stream().filter(it -> it.getSimpleName().equals(m.getName()))
-                    .findFirst().orElseThrow(AssertionError::new);
-            
-            assertMethodMatch(m, mSym);
-        }
-    }
-    
-    private void assertMethodMatch(Method m, JMethodSymbol mSym) {
-        Assert.assertEquals(m.getParameterCount(), mSym.getArity());
-        
-        final Parameter[] parameters = m.getParameters();
-        List<JFormalParamSymbol> formals = mSym.getFormalParameters();
-        
-        for (int i = 0; i < m.getParameterCount(); i++) {
-            assertParameterMatch(parameters[i], formals.get(i));
-        }
-        
-        // Defaults should match too (even if not an annotation method, both should be null)
-        Assert.assertEquals(symValueOf(m.getDefaultValue()), mSym.getDefaultAnnotationValue());
-    }
-
-    private void assertParameterMatch(Parameter p, JFormalParamSymbol pSym) {
-        if (expectDebugSymbols) {
-            if (p.isNamePresent()) {
-                Assert.assertEquals(p.getName(), pSym.getSimpleName());
-                Assert.assertEquals(Modifier.isFinal(p.getModifiers()), pSym.isFinal());
-            } else {
-                System.out.println("WARN: test classes were not compiled with -parameters, parameters not fully checked");
-            }
-        } else {
-            // note that this asserts, that the param names are unavailable
-            Assert.assertEquals("", pSym.getSimpleName());
-            Assert.assertFalse(pSym.isFinal());
-        }
-        
-        // Ensure type matches
-        final JTypeMirror expectedType = typeMirrorOf(p.getType());
-        
-        Assert.assertEquals(expectedType, pSym.getTypeMirror(Substitution.EMPTY));
-    }
-    
     private void assertEqualsAnnotations(AnnotatedElement e, AnnotableSymbol eSym) {
         Assert.assertEquals(e.getDeclaredAnnotations().length, eSym.getDeclaredAnnotations().size());
         for (Annotation annot : e.getDeclaredAnnotations()) {
@@ -284,12 +233,5 @@ public abstract class AbstractSymbolTest {
             }
         }
     }
-    
-    private JTypeMirror typeMirrorOf(Class<?> type) {
-        if (type.isPrimitive()) {
-            return ts.getPrimitive(PrimitiveTypeKind.fromName(type.getName()));
-        }
-        
-        return ts.typeOf(loader.resolveClassFromBinaryName(type.getName()), true);
-    }
+
 }
