+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symbols/SymbolicValueTest.java
@@ -1,3 +1,6 @@
+/*
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
 package net.sourceforge.pmd.lang.java.symbols;
 
 import java.util.Arrays;
